requirements.txt
azure-storage-file-share
python-dotenv
flask